backend/requirements.txt
fastapi==0.75.0
fastapi_mail==1.0.9
email-validator==1.1.3
sqlalchemy==1.4.31
alembic==1.7.6
mysqlclient==2.1.0
uvicorn==0.17.5
python-jose==3.3.0
passlib==1.7.4
